+++ app/Livewire/Admin/Documents/DeleteDocument.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Admin\Documents;
 
 use App\Models\Admin\Document;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use LivewireUI\Modal\ModalComponent;
 
@@ -17,10 +18,13 @@ class DeleteDocument extends ModalComponent
 
     public function deleteDocument()
     {
+        $document = Document::findOrFail($this->documentId);
+        // Verifica nuovamente l'autorizzazione
+        if (!Auth::check() || $document->admin_id !== Auth::id()) {
+            abort(403, 'You are not authorized');
+        }
+
         try {
-
-            $document = Document::findOrFail($this->documentId);
-
             if ($document) {
                 Storage::disk('public')->delete($document->img_url);
                 $document->delete();
@@ -29,7 +33,6 @@ class DeleteDocument extends ModalComponent
             $this->dispatch('closeModal');
             session()->flash('message', 'Document deleted successfully!');
             $this->redirect('/admin/documents', navigate: true);
-
         } catch (\Illuminate\Validation\ValidationException $e) {
             $this->dispatch('validation-error', $e->validator->errors()->first());
             return;
