+++ app/Livewire/Admin/Projects/DeleteProject.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Admin\Projects;
 
 use App\Models\Admin\Project;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use LivewireUI\Modal\ModalComponent;
 
@@ -17,10 +18,14 @@ class DeleteProject extends ModalComponent
 
     public function deleteProject()
     {
+        $project = Project::findOrFail($this->projectId);
+        
+        // Verifica nuovamente l'autorizzazione
+        if (!Auth::check() || $project->admin_id !== Auth::id()) {
+            abort(403, 'You are not authorized');
+        }
+
         try {
-
-            $project = Project::findOrFail($this->projectId);
-
             if ($project) {
                 Storage::disk('public')->delete($project->img_url);
                 $project->delete();
@@ -29,7 +34,6 @@ class DeleteProject extends ModalComponent
             $this->dispatch('closeModal');
             session()->flash('message', 'Project deleted successfully!');
             $this->redirect('/admin/projects', navigate: true);
-
         } catch (\Illuminate\Validation\ValidationException $e) {
             $this->dispatch('validation-error', $e->validator->errors()->first());
             return;
